requirements.txt
chardet==3.0.4
cssselect==1.0.1
cycler==0.10.0
Cython==0.27.3
Genshi==0.7
lxml==4.1.1
Manifold==0.2.0
matplotlib==2.1.1
numpy==1.13.3
pdfminer.six==20170720
pdfquery==0.4.3
pycryptodome==3.4.7
pyparsing==2.2.0
pyPdf==1.13
PyPDF2==1.26.0
pyquery==1.3.0
python-dateutil==2.6.1
pytz==2017.3
roman==2.0.0
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
urllib5==5.0.0
